--- trello-api/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "trello-ts-api",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "axios": "^1.13.6",
-    "dotenv": "^17.3.1"
-  },
-  "devDependencies": {
-    "@types/node": "^25.3.5",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.9.3"
-  }
-}
